Manual combo-layer update from meta-security-isafw

+++ meta-security-isafw/lib/isafw/isaplugins/ISA_cve_plugin.py
@@ -32,30 +32,34 @@ import re
 import tempfile
 
 CVEChecker = None
-cve_report = "/cve-report"
 pkglist = "/cve_check_tool_pkglist"
-log = "/isafw_cvelog"
 
 class ISA_CVEChecker:    
     initialized = False
     def __init__(self, ISA_config):
         self.proxy = ISA_config.proxy
         self.reportdir = ISA_config.reportdir
-        self.logdir = ISA_config.logdir
         self.timestamp = ISA_config.timestamp
+        self.logfile = ISA_config.logdir + "/isafw_cvelog"
+        self.report_name = ISA_config.reportdir + "/cve_report_" + ISA_config.machine + "_" + ISA_config.timestamp  
+        output = ""
         # check that cve-check-tool is installed
-        rc = subprocess.call(["which", "cve-check-tool"])
-        if rc == 0:
-            self.initialized = True
-            print("Plugin ISA_CVEChecker initialized!")
-            with open(self.logdir + log, 'a') as flog:
-                flog.write("\nPlugin ISA_CVEChecker initialized!\n")
+        try:
+            popen = subprocess.Popen("which cve-check-tool", shell=True, stdout=subprocess.PIPE)
+            popen.wait()
+            output = popen.stdout.read()
+        except:
+            with open(self.logfile, 'a') as flog:
+                flog.write("error executing which cve-check-tool\n")
         else:
-            print("cve-check-tool is missing!")
-            print("Please install it from https://github.com/ikeydoherty/cve-check-tool.")
-            with open(self.logdir + log, 'a') as flog:
-                flog.write("cve-check-tool is missing!\n")
-                flog.write("Please install it from https://github.com/ikeydoherty/cve-check-tool.\n")
+            if output:
+                self.initialized = True
+                with open(self.logfile, 'a') as flog:
+                    flog.write("\nPlugin ISA_CVEChecker initialized!\n")
+            else:
+                with open(self.logfile, 'a') as flog:
+                    flog.write("cve-check-tool is missing!\n")
+                    flog.write("Please install it from https://github.com/ikeydoherty/cve-check-tool.\n")
 
     def process_package(self, ISA_pkg):
         if (self.initialized == True):
@@ -73,45 +77,47 @@ class ISA_CVEChecker:
                     for a in alias_pkgs_faux:
                         fauxfile.write(a)
 
-                with open(self.logdir + log, 'a') as flog:
+                with open(self.logfile, 'a') as flog:
                     flog.write("\npkg info: " + pkgline_faux)
             else:
-                print("Mandatory arguments such as pkg name, version and list of patches are not provided!")
-                print("Not performing the call.")
                 self.initialized = False
-                with open(self.logdir + log, 'a') as flog:
+                with open(self.logfile, 'a') as flog:
                     flog.write("Mandatory arguments such as pkg name, version and list of patches are not provided!\n")
                     flog.write("Not performing the call.\n")
         else:
-            print("Plugin hasn't initialized! Not performing the call.")
-            with open(self.logdir + log, 'a') as flog:
+            with open(self.logfile, 'a') as flog:
                 flog.write("Plugin hasn't initialized! Not performing the call.\n")
 
     def process_report(self):
+        if not os.path.isfile(self.reportdir + pkglist + "_" + self.timestamp + ".faux"): 
+            return
         if (self.initialized == True):
-            print("Creating report in HTML format.")
-            with open(self.logdir + log, 'a') as flog:
+            with open(self.logfile, 'a') as flog:
                 flog.write("Creating report in HTML format.\n")
             self.process_report_type("html")
 
-            print("Creating report in CSV format.")
-            with open(self.logdir + log, 'a') as flog:
+            with open(self.logfile, 'a') as flog:
                 flog.write("Creating report in CSV format.\n")
             self.process_report_type("csv")
 
             pkglist_faux = pkglist + "_" + self.timestamp + ".faux"
             os.remove(self.reportdir + pkglist_faux)
 
-            print("Creating report in XML format.")
-            with open(self.logdir + log, 'a') as flog:
+            with open(self.logfile, 'a') as flog:
                 flog.write("Creating report in XML format.\n")
             self.write_report_xml()
 
     def write_report_xml(self):
-        from lxml import etree
+        try:
+            from lxml import etree
+        except ImportError:
+            try:
+                import xml.etree.cElementTree as etree
+            except ImportError:
+                import xml.etree.ElementTree as etree
         numTests = 0
         root = etree.Element('testsuite', name='CVE_Plugin', tests='1')
-        with open(self.reportdir + cve_report + "_" + self.timestamp + ".csv", 'r') as f:
+        with open(self.report_name + ".csv", 'r') as f:
             for line in f:
                 numTests += 1
                 line = line.strip()
@@ -122,8 +128,12 @@ class ISA_CVEChecker:
                     tcase = etree.SubElement(root, 'testcase', classname='ISA_CVEChecker', name=line.split(',',1)[0])
         root.set('tests', str(numTests))
         tree = etree.ElementTree(root)
-        output = self.reportdir +  cve_report + "_" + self.timestamp + '.xml' 
-        tree.write(output, encoding= 'UTF-8', pretty_print=True, xml_declaration=True)
+        output = self.report_name + '.xml' 
+        try:
+            tree.write(output, encoding='UTF-8', pretty_print=True, xml_declaration=True)
+        except TypeError:
+            tree.write(output, encoding='UTF-8', xml_declaration=True)
+
 
     def process_report_type(self, rtype):
         # now faux file is ready and we can process it
@@ -136,19 +146,19 @@ class ISA_CVEChecker:
             rtype = "csv"
         pkglist_faux = pkglist + "_" + self.timestamp + ".faux"
         args += "-a -t faux '" + self.reportdir + pkglist_faux  + "'"
+        with open(self.logfile, 'a') as flog:
+            flog.write("Args: " + args)
         try:
-            popen = subprocess.Popen(args, shell=True, stdout=subprocess.PIPE)
-            popen.wait()
-            output = popen.stdout.read()
+            popen = subprocess.Popen(args, shell=True, stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
+            stdout_value = popen.communicate()[0]
         except:
-            print("Error in executing cve-check-tool: ", sys.exc_info())
-            output = "Error in executing cve-check-tool"
-            with open(self.logdir + log, 'a') as flog:
+            stdout_value = "Error in executing cve-check-tool"
+            with open(self.logfile, 'a') as flog:
                 flog.write("Error in executing cve-check-tool: " + sys.exc_info())
         else:
-            report = cve_report + "_" + self.timestamp + "." + rtype
-            with open(self.reportdir + report, 'w') as freport:
-                freport.write(output)
+            report = self.report_name + "." + rtype
+            with open(report, 'w') as freport:
+                freport.write(stdout_value)
 
     def process_patch_list(self, patch_files):
         patch_info = ""
